Azure Database is now Reachable and rest Funktion are right

+++ DnssWebApi/Dto/AdmaModelSendingJustValueDto.cs
@@ -7,6 +7,6 @@ namespace DnssWebApi.Dto
 {
     public class AdmaModelSendingJustValueDto
     {
-        public float Value { get; set; }
+        public int Value { get; set; }
     }
 }
